Change contents of getDescription from a formatted line
Previously, the getDescription method could return both the 0th and the 1st field of the formatted line depending on the situation. Now it only returns the 1st field aka the real description. Introduced getDescriptionCode to return the code in the 0th field.

Also added some files to the CMakeLists.txt to make them show up in Code::Blocks.