Change how the result is written
The main script file will now directly contain the damage output of the last command.
It wasn't possible to execute more than one script from another script so it had to be done this way.

Try . as a fill character as spaces in AO do not have the same size as other chars.

Move the percentage sign from heading to behind the nr.