Several fixes to Damage writing.
* Calculate DPM on dealt and received on every getDamage in Player.
Can't know which one will be used before hand.
* Split getTotalDamageForEachPlayer into two methods. One for damage
dealt to the player and one for received from the player. Done to avoid
getting players that have not for example dealt any damage when we want
to write damage dealt.
* Fix crit/miss calculation. There is still problems with this as it
should not include special and shield damage (which is counted as
regular damage) but it's slightly better now.